// src/api/mail/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface MailApi {
	mailSend: (request: Api.MailSendMailRequest) => Promise<MailSendResponse>
	mailVerify: (request: Api.MailVerifyMailRequest) => Promise<MailVerifyResponse>
}

export type MailSendResponse = MailSend200Response | MailSendDefaultResponse

export interface MailSend200Response {
	status: 200
	body: Api.MailSendMailResponse
	headers?: never
}

export interface MailSendDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type MailVerifyResponse = MailVerify200Response | MailVerifyDefaultResponse

export interface MailVerify200Response {
	status: 200
	body: Api.MailVerifyMailResponse
	headers?: never
}

export interface MailVerifyDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}